chat realTime

+++ src/database/migrations/2023_09_09_192039_create_messages_table.php
@@ -13,8 +13,7 @@ return new class extends Migration
     {
         Schema::create('messages', function (Blueprint $table) {
             $table->id();
-         
-            $table->int('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->string('content');
             $table->timestamps();
         });
